Added better unit tests for Quboles Hook and Operator
