+++ tests/providers/qubole/hooks/test_qubole.py
@@ -16,14 +16,28 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-import unittest
+from unittest import TestCase, mock
+from qds_sdk.commands import PrestoCommand
 
 from airflow.providers.qubole.hooks.qubole import QuboleHook
+
+DAG_ID = "qubole_test_dag"
+TASK_ID = "test_task"
+RESULTS_WITH_HEADER = 'header1\theader2\nval1\tval2'
+RESULTS_WITH_NO_HEADER = 'val1\tval2'
 
 add_tags = QuboleHook._add_tags
 
 
-class TestQuboleHook(unittest.TestCase):
+# pylint: disable =unused-arguments
+def get_result_mock(fp, inline, delim, fetch, arguments):
+    if arguments[0] == 'true':
+        fp.write(bytearray(RESULTS_WITH_HEADER, 'utf-8'))
+    else:
+        fp.write(bytearray(RESULTS_WITH_NO_HEADER, 'utf-8'))
+
+
+class TestQuboleHook(TestCase):
     def test_add_string_to_tags(self):
         tags = {'dag_id', 'task_id'}
         add_tags(tags, 'string')
@@ -38,3 +52,28 @@ class TestQuboleHook(unittest.TestCase):
         tags = {'dag_id', 'task_id'}
         add_tags(tags, ('value1', 'value2'))
         assert {'dag_id', 'task_id', 'value1', 'value2'} == tags
+
+    @mock.patch('qds_sdk.commands.Command.get_results', new=get_result_mock)
+    def test_get_results_with_headers(self):
+        dag = mock.MagicMock()
+        dag.dag_id = DAG_ID
+        hook = QuboleHook(task_id=TASK_ID, command_type='prestocmd', dag=dag)
+
+        task = mock.MagicMock()
+        task.xcom_pull.return_value = 'test_command_id'
+        with mock.patch('qds_sdk.resource.Resource.find', return_value=PrestoCommand):
+            results = open(hook.get_results(ti=task, include_headers=True), 'r').read()
+            assert results == RESULTS_WITH_HEADER
+
+    @mock.patch('qds_sdk.commands.Command.get_results', new=get_result_mock)
+    def test_get_results_without_headers(self):
+        dag = mock.MagicMock()
+        dag.dag_id = DAG_ID
+        hook = QuboleHook(task_id=TASK_ID, command_type='prestocmd', dag=dag)
+
+        task = mock.MagicMock()
+        task.xcom_pull.return_value = 'test_command_id'
+
+        with mock.patch('qds_sdk.resource.Resource.find', return_value=PrestoCommand):
+            results = open(hook.get_results(ti=task, include_headers=False), 'r').read()
+            assert results == RESULTS_WITH_NO_HEADER
