+++ tests/providers/qubole/operators/test_qubole.py
@@ -17,9 +17,7 @@
 # under the License.
 #
 
-from unittest import mock, TestCase
-
-from qds_sdk.commands import PrestoCommand
+from unittest import TestCase, mock
 
 from airflow import settings
 from airflow.models import DAG, Connection
@@ -36,16 +34,6 @@ DEFAULT_CONN = "qubole_default"
 TEMPLATE_CONN = "my_conn_id"
 TEST_CONN = "qubole_test_conn"
 DEFAULT_DATE = datetime(2017, 1, 1)
-
-RESULTS_WITH_HEADER = 'header1\theader2\nval1\tval2'
-RESULTS_WITH_NO_HEADER = 'val1\tval2'
-
-
-def get_result_mock(fp, arguments):
-    if arguments[0] == 'true':
-        fp.write(bytearray(RESULTS_WITH_HEADER, 'utf-8'))
-    else:
-        fp.write(bytearray(RESULTS_WITH_NO_HEADER, 'utf-8'))
 
 
 class TestQuboleOperator(TestCase):
@@ -193,26 +181,16 @@ class TestQuboleOperator(TestCase):
         op = QuboleOperator(task_id=TASK_ID, pool=test_pool)
         assert op.pool == test_pool
 
-    @mock.patch('qds_sdk.commands.Command.get_results', new=get_result_mock)
-    def test_get_results_with_headers(self):
+    @mock.patch('airflow.providers.qubole.hooks.qubole.QuboleHook.get_results')
+    def test_parameter_include_header_passed(self, mock_get_results):
         dag = DAG(DAG_ID, start_date=DEFAULT_DATE)
+        qo = QuboleOperator(task_id=TASK_ID, dag=dag, command_type='prestocmd')
+        qo.get_results(include_headers=True)
+        mock_get_results.asset_called_with('include_headers', True)
 
-        task = QuboleOperator(task_id=TASK_ID, command_type='prestocmd', dag=dag)
-
-        with mock.patch.object(task, 'xcom_pull', return_value='test_command_id'):
-            with mock.patch('qds_sdk.resource.Resource.find', return_value=PrestoCommand):
-                results = open(task.get_results(ti=task, include_headers=True), 'r').read()
-                assert results == RESULTS_WITH_HEADER
-                results = open(task.get_results(ti=task, include_headers=False), 'r').read()
-                assert results == RESULTS_WITH_NO_HEADER
-
-    @mock.patch('qds_sdk.commands.Command.get_results', new=get_result_mock)
-    def test_get_results_without_headers(self):
+    @mock.patch('airflow.providers.qubole.hooks.qubole.QuboleHook.get_results')
+    def test_parameter_include_header_missing(self, mock_get_results):
         dag = DAG(DAG_ID, start_date=DEFAULT_DATE)
-
-        task = QuboleOperator(task_id=TASK_ID, command_type='prestocmd', dag=dag)
-
-        with mock.patch.object(task, 'xcom_pull', return_value='test_command_id'):
-            with mock.patch('qds_sdk.resource.Resource.find', return_value=PrestoCommand):
-                results = open(task.get_results(ti=task, include_headers=False), 'r').read()
-                assert results == RESULTS_WITH_NO_HEADER
+        qo = QuboleOperator(task_id=TASK_ID, dag=dag, command_type='prestocmd')
+        qo.get_results()
+        mock_get_results.asset_called_with('include_headers', False)
